Removed cyclic dependency in ingestion
